+++ src/presets/index.ts
@@ -1,3 +1,4 @@
 import ableism from "./ableism.json";
+import condescension from "./condescension.json";
 
-export default { ableism };
+export default { ableism, condescension };